refactored endpoinnts to be email ready
