+++ src/main/kotlin/pt/up/fe/ni/website/backend/controller/AuthController.kt
@@ -1,5 +1,6 @@
 package pt.up.fe.ni.website.backend.controller
 
+import jakarta.validation.Valid
 import org.springframework.beans.factory.annotation.Value
 import org.springframework.security.access.prepost.PreAuthorize
 import org.springframework.web.bind.annotation.GetMapping
@@ -9,6 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import pt.up.fe.ni.website.backend.dto.auth.LoginDto
+import pt.up.fe.ni.website.backend.dto.auth.PasswordRecoveryConfirmDto
+import pt.up.fe.ni.website.backend.dto.auth.PasswordRecoveryRequestDto
 import pt.up.fe.ni.website.backend.dto.auth.TokenDto
 import pt.up.fe.ni.website.backend.model.Account
 import pt.up.fe.ni.website.backend.service.AuthService
@@ -33,12 +36,19 @@ class AuthController(val authService: AuthService) {
         return mapOf("access_token" to accessToken)
     }
 
-    @PostMapping("/recoverPassword/{id}")
-    fun generateRecoveryToken(@PathVariable id: Long): Map<String, String> {
-        val recoveryToken = authService.generateRecoveryToken(id)
+    @PostMapping("/password/recovery")
+    fun generateRecoveryToken(@RequestBody recoveryRequestDto: PasswordRecoveryRequestDto): Map<String, String> {
+        val recoveryToken = authService.generateRecoveryToken(recoveryRequestDto.email)
         // TODO: Change to email service
-        return mapOf("recovery_url" to "$recoverPasswordPage/$recoveryToken")
+        return mapOf("recovery_url" to "$recoverPasswordPage/$recoveryToken/confirm")
     }
+
+    @PostMapping("/password/recovery/{token}/confirm")
+    fun confirmRecoveryToken(
+        @Valid @RequestBody
+        dto: PasswordRecoveryConfirmDto,
+        @PathVariable token: String
+    ) = authService.confirmRecoveryToken(token, dto)
 
     @GetMapping
     @PreAuthorize("hasRole('MEMBER')")
