+++ src/main/resources/application.properties
@@ -39,7 +39,7 @@ upload.static-serve=http://localhost:3000/static
 
 # Recover password page
 # (for now it's the backend endpoint as we don't have the front end page yet)
-page.recover-password=http://localhost:8080/accounts/recoverPassword
+page.recover-password=http://localhost:8080/auth/password/recovery
 
 # Cors Origin
 cors.allow-origin = http://localhost:3000
